requirements.txt
numpy
pyannote.audio
cachetools>=2.0.0
librosa
pandas>=0.18.0
pyYAML>=3.12
scikit-learn>=0.20.2
sortedcontainers>=2.0.4
soundfile>=0.10.2
tqdm>=4.29.1
tensorboard>=2.0.0
typing_extensions>=3.7.4
torch>=1.4.0
tensorboard>=2.0.0
pyannote.core
pyannote.database
pyannote.pipeline==1.5
git+ssh://git@github.com/pyannote/pyannote-pipeline@develop#egg=pyannote.pipeline

